refactor: move proposals to a folder
